backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
passlib[bcrypt]
python-jose[cryptography]
celery[redis]
redis
aiofiles
motor
pymongo[srv]
pyjwt
websockets
httpx
python-multipart
shapely
web3
paho-mqtt
pyshp
fastkml
numpy
opencv-python-headless
Pillow
reportlab
boto3
rasterio
scikit-image
